Work around dynamic method definition
This is done in two steps:

- Split per-version code, each into their own file
- Type only for one (current) version category
- Ignore the other (legacy) type category

Both could be type-checked with targets, but this would need this issue
to be solved: https://github.com/soutaro/steep/issues/978
