+++ lib/graft/callback/callable.ruby2.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Graft
+  class Callback
+    module Callable
+      def call(*args, &block)
+        return unless enabled?
+
+        @block.call(*args, &block)
+      end
+    end
+  end
+end
