+++ lib/graft/callback/callable.ruby3.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Graft
+  class Callback
+    module Callable
+      def call(*args, **kwargs, &block)
+        return unless enabled?
+
+        @block.call(*args, **kwargs, &block)
+      end
+    end
+  end
+end
