+++ lib/graft/hook_point.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
 
+require_relative "hook_point/prepend"
+require_relative "hook_point/chain"
+
 module Graft
   class HookPointError < StandardError; end
 
@@ -158,256 +161,6 @@ module Graft
 
     def disabled?(key) # rubocop:disable Lint/UselessMethodDefinition
       super
-    end
-
-    module Prepend
-      def installed?(key)
-        prepended?(key) && overridden?(key)
-      end
-
-      def install(key, &block)
-        prepend(key)
-        override(key, &block)
-      end
-
-      def uninstall(key)
-        unoverride(key) if overridden?(key)
-      end
-
-      def enable(key)
-        raise HookPointError, "enable(#{key.inspect}) with prepend strategy"
-      end
-
-      def disable(key)
-        unoverride(key)
-      end
-
-      def disabled?(key)
-        !overridden?(key)
-      end
-
-      private
-
-      def hook_module(key)
-        target = klass_method? ? klass.singleton_class : klass
-
-        mod = target.ancestors.each do |e|
-          break if e == target
-          break(e) if e.instance_of?(HookModule) && e.key == key
-        end
-
-        raise HookPointError, "Inconsistency detected: #{target} missing from its own ancestors" if mod.is_a?(Array)
-
-        [target, mod]
-      end
-
-      def prepend(key)
-        target, mod = hook_module(key)
-
-        mod ||= HookModule.new(key)
-
-        target.prepend(mod)
-      end
-
-      def prepended?(key)
-        _, mod = hook_module(key)
-
-        mod != nil
-      end
-
-      def overridden?(key)
-        _, mod = hook_module(key)
-
-        (
-          mod.instance_methods(false) +
-          mod.protected_instance_methods(false) +
-          mod.private_instance_methods(false)
-        ).include?(method_name)
-      end
-
-      def override(key, &block)
-        hook_point = self
-        method_name = @method_name
-
-        _, mod = hook_module(key)
-
-        mod.instance_eval do
-          if hook_point.private_method?
-            private
-          elsif hook_point.protected_method?
-            protected
-          else
-            public
-          end
-
-          define_method(:"#{method_name}", &block)
-        end
-      end
-
-      def unoverride(key)
-        method_name = @method_name
-
-        _, mod = hook_module(key)
-
-        mod.instance_eval { remove_method(method_name) }
-      end
-    end
-
-    module Chain
-      def installed?(key)
-        defined(key)
-      end
-
-      def install(key, &block)
-        define(key, &block)
-        chain(key)
-      end
-
-      def uninstall(key)
-        disable(key)
-        remove(key)
-      end
-
-      def enable(key)
-        chain(key)
-      end
-
-      def disable(key)
-        unchain(key)
-      end
-
-      def disabled?(key)
-        !chained?(key)
-      end
-
-      private
-
-      def defined(suffix)
-        if klass_method?
-          (
-            klass.methods +
-            klass.protected_methods +
-            klass.private_methods
-          ).include?(:"#{method_name}_with_#{suffix}")
-        elsif instance_method?
-          (
-            klass.instance_methods +
-            klass.protected_instance_methods +
-            klass.private_instance_methods
-          ).include?(:"#{method_name}_with_#{suffix}")
-        else
-          raise HookPointError, "#{self} unknown hook point kind"
-        end
-      end
-
-      def define(suffix, &block)
-        hook_point = self
-        method_name = @method_name
-
-        if klass_method?
-          klass.singleton_class.instance_eval do
-            if hook_point.private_method?
-              private
-            elsif hook_point.protected_method?
-              protected
-            else
-              public
-            end
-
-            define_method(:"#{method_name}_with_#{suffix}", &block)
-          end
-        elsif instance_method?
-          klass.class_eval do
-            if hook_point.private_method?
-              private
-            elsif hook_point.protected_method?
-              protected
-            else
-              public
-            end
-
-            define_method(:"#{method_name}_with_#{suffix}", &block)
-          end
-        else
-          raise HookPointError, "unknown hook point kind"
-        end
-      end
-
-      def remove(suffix)
-        method_name = @method_name
-
-        if klass_method?
-          klass.singleton_class.instance_eval do
-            remove_method(:"#{method_name}_with_#{suffix}")
-          end
-        elsif instance_method?
-          klass.class_eval do
-            remove_method(:"#{method_name}_with_#{suffix}")
-          end
-        else
-          raise HookPointError, "unknown hook point kind"
-        end
-      end
-
-      def chained?(suffix)
-        method_name = @method_name
-
-        if klass_method?
-          klass.singleton_class.instance_eval do
-            instance_method(:"#{method_name}").original_name == :"#{method_name}_with_#{suffix}"
-          end
-        elsif instance_method?
-          klass.class_eval do
-            instance_method(:"#{method_name}").original_name == :"#{method_name}_with_#{suffix}"
-          end
-        else
-          raise HookPointError, "unknown hook point kind"
-        end
-      end
-
-      def chain(suffix)
-        method_name = @method_name
-
-        if klass_method?
-          klass.singleton_class.instance_eval do
-            alias_method :"#{method_name}_without_#{suffix}", :"#{method_name}"
-            alias_method :"#{method_name}", :"#{method_name}_with_#{suffix}"
-          end
-        elsif instance_method?
-          klass.class_eval do
-            alias_method :"#{method_name}_without_#{suffix}", :"#{method_name}"
-            alias_method :"#{method_name}", :"#{method_name}_with_#{suffix}"
-          end
-        else
-          raise HookPointError, "unknown hook point kind"
-        end
-      end
-
-      def unchain(suffix)
-        method_name = @method_name
-
-        if klass_method?
-          klass.singleton_class.instance_eval do
-            alias_method :"#{method_name}", :"#{method_name}_without_#{suffix}"
-          end
-        elsif instance_method?
-          klass.class_eval do
-            alias_method :"#{method_name}", :"#{method_name}_without_#{suffix}"
-          end
-        end
-      end
-
-      # @dynamic apply
-      # Using `define_method` instead of `def` as the latter trips up static type checking
-      if RUBY_VERSION < "3.0"
-        define_method :apply do |obj, suffix, *args, &block|
-          obj.send(:"#{method_name}_without_#{suffix}", *args, &block)
-        end
-      else
-        define_method :apply do |obj, suffix, *args, **kwargs, &block|
-          obj.send(:"#{method_name}_without_#{suffix}", *args, **kwargs, &block)
-        end
-      end
     end
   end
 end
