+++ lib/graft/hook_point/chain/applyable.ruby3.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Graft
+  class HookPoint
+    module Chain
+      module Applyable
+        def apply(obj, suffix, *args, **kwargs, &block)
+          obj.send(:"#{method_name}_without_#{suffix}", *args, **kwargs, &block)
+        end
+      end
+    end
+  end
+end
