+++ sig/graft/callback/callable.rbs
@@ -1,0 +1,8 @@
+module Graft
+  class Callback
+    module Callable: Callback
+    # def call: (*untyped args) { () -> untyped } -> (nil | untyped)
+      def call: (*untyped args, **untyped kwargs) { () -> untyped } -> (nil | untyped)
+    end
+  end
+end
