+++ sig/graft/hook_point.rbs
@@ -65,66 +65,5 @@ module Graft
     def disable: (String key) -> void
 
     def disabled?: (String key) -> bool
-
-    module Prepend : HookPoint
-      def installed?: (String key) -> bool
-
-      def install: (String key) { () -> untyped } -> void
-
-      def uninstall: (String key) -> void
-
-      def enable: (String key) -> void
-
-      def disable: (String key) -> void
-
-      def disabled?: (String key) -> bool
-
-      private
-
-      def hook_module: (String key) -> ::Array[untyped]
-
-      def prepend: (String key) -> void
-
-      def prepended?: (String key) -> bool
-
-      def overridden?: (String key) -> bool
-
-      def override: (String key) { () -> untyped } -> void
-
-      def unoverride: (String key) -> void
-    end
-
-    module Chain : HookPoint
-      def installed?: (String key) -> bool
-
-      def install: (String key) { () -> untyped } -> void
-
-      def uninstall: (String key) -> void
-
-      def enable: (String key) -> void
-
-      def disable: (String key) -> void
-
-      def disabled?: (String key) -> bool
-
-      private
-
-      def defined: (untyped suffix) -> untyped
-
-      def define: (untyped suffix) { () -> untyped } -> void
-
-      def remove: (untyped suffix) -> void
-
-      def chained?: (untyped suffix) -> bool
-
-      def chain: (untyped suffix) -> void
-
-      def unchain: (untyped suffix) -> void
-
-    # TODO: resolve version specific dynamism
-
-    # def apply: (untyped obj, untyped suffix, *untyped args) { () -> untyped } -> untyped
-      def apply: (untyped obj, untyped suffix, *untyped args, **untyped kwargs) { () -> untyped } -> untyped
-    end
   end
 end
