+++ sig/graft/hook_point/prepend.rbs
@@ -1,0 +1,31 @@
+module Graft
+  class HookPoint
+    module Prepend : HookPoint
+      def installed?: (String key) -> bool
+
+      def install: (String key) { () -> untyped } -> void
+
+      def uninstall: (String key) -> void
+
+      def enable: (String key) -> void
+
+      def disable: (String key) -> void
+
+      def disabled?: (String key) -> bool
+
+      private
+
+      def hook_module: (String key) -> ::Array[untyped]
+
+      def prepend: (String key) -> void
+
+      def prepended?: (String key) -> bool
+
+      def overridden?: (String key) -> bool
+
+      def override: (String key) { () -> untyped } -> void
+
+      def unoverride: (String key) -> void
+    end
+  end
+end
